Macro akin::akin

akin!() { /* proc-macro */ }

Duplicates the given code and substitutes specific identifiers for different code snippets in each duplicate.

Usage

Write each identifier following let &ident = [v1, v2, v3, ...], and use them in the snippet you want to duplicate with *ident.

Code snippets are copied max(used_vars.values) times.

akin! {
    let &var = ['a', 'b'];
    println!("{}", *var);
}

Will get copied 2 times, because the variable &var has 2 values.

If a used variable has less values than another, the last one will be used.

akin! {
    let &v1 = [c];
    let &v2 = [a, b];
    println!("*v1*v2");
}

Expands to

println!("ca");
println!("cb");

All code in variables must be enclosed in brackets {...}.

akin! {
    let &var = [-1, 1];
    let &code = [
        {
            println!("true");
        },
        {
            println!("false");
        }
    ];

    if *var < 0 {
        *code
    }
}

Will expand to

if -1 < 0 {
    println!("true");
}
if 1 < 0 {
    println!("false")
}

Example

akin! {
    let &a = [1, 2, 3, 4, 5, 6];
    let &b = [4, 5, 6];
    let &code = {
        println!("*a + *b = {}", *a + *b);
    };
    let print = true;
    if print {
        *code
    }
}

Expands to

let print = true;
if print {
   println!("1 + 4 = 5");
   println!("2 + 5 = 7");
   println!("3 + 6 = 9");
   println!("4 + 6 = 10");
   println!("5 + 6 = 11");
   println!("6 + 6 = 12");
}

Range syntax

As a matter of convenience, the range syntax is also accepted, when declaring a variable, e.g. 0..3 and 0..=3, which are equivalent to [0,1,2] and [0,1,2,3] respectively. So the variables a and b in the example above could also be declared like

let &a = 1..=6;
let &b = 4..=6;

Presently, only unsigned integers that can fit in u64 are supported in ranges, i.e. ranges like -10..-1 or 'a'..'c', which are fine in regular Rust, aren’t accepted by akin.

NONE

NONE is the way you can tell akin to simply skip that value and not write anything. It is useful for when you want to have elements in a duplication that do not have to be in the others.

akin! {
    let &num = [1, 2, 3];
    let &code = [
        NONE,
        {
            .pow(2)
        }
    ];

    println!("*num^2 = {}", *num~u32*code);
    // *num~u32 is necessary to ensure the type is written correctly (it would be "1 u32" without it)
}

Joint modifier

By default, akin places a space between all identifiers. Sometimes, this is not desirable, for example, if trying to interpolate between a function name

akin! {
    let &name = [1];
    fn _*name()
}

Will get wrongly expanded because ‘_’ is an identifier

fn _ 1()

To avoid it, use the joint modifier ~, making the next identifier not to be separated.

akin! {
    let &name = [1];
    fn _~*name() // *name is affected by the modifier
}

Will get correctly expanded to

fn _1()

Inside string literals "..." it is not necessary to use the modifier, as Rust does not count them as identifiers.

This is a limitation on proc_macro parsing, so I doubt it’ll be fixed soon.

More examples

trait Sqrt {
    fn dumb_sqrt(self) -> Result<f64, &'static str>;
}

akin! {
    let &int_type = [i64, u64];
    let &negative_check = [
        {
            if self < 0 {
                return Err("Sqrt of negative number")
            }
        },
        NONE
    ];

    let &num = [1,     2,    3,  4];
    let &res = [1., 1.41, 1.73,  2.];
    let &branch = {
        *num => Ok(*res),
    };

    impl Sqrt for *int_type {
        fn dumb_sqrt(self) -> Result<f64, &'static str> {
            *negative_check

            match self {
                *branch
                _ => Err("Sqrt of num not in [1, 4]")
            }
        }
    }
}

Is expanded to:

trait Sqrt {
    fn dumb_sqrt(self) -> Result<f64, &'static str>;
}

impl Sqrt for i64 {
    fn dumb_sqrt(self) -> Result<f64, &'static str> {
        if self < 0 {
            return Err("Sqrt of negative number")
        }

        match self {
            1 => Ok(1.),
            2 => Ok(1.41),
            3 => Ok(1.73),
            4 => Ok(2.),
            _ => Err("Sqrt of num not in [1, 4]")
        }
    }
}

impl Sqrt for u64 {
    fn dumb_sqrt(self) -> Result<f64, &'static str> {
        match self {
            1 => Ok(1.),
            2 => Ok(1.41),
            3 => Ok(1.73),
            4 => Ok(2.),
            _ => Err("Sqrt of num not in [1, 4]")
        }
    }
}